fix: create custom blog curd

+++ src/modules/superadmin/amenityOption/amenity-option.resolver.ts
@@ -1,9 +1,17 @@
-import { AmenityOption, AmenityOptionList } from '@src/models/global.model';
+import {
+  AmenityOption,
+  AmenityOptionList,
+  CtxType,
+} from '@src/models/global.model';
 import { AmenityOptionService } from './amenity-option.services';
-import { Resolver, Query, Mutation, Args, Int } from '@nestjs/graphql';
+import { Resolver, Query, Mutation, Args, Int, Context } from '@nestjs/graphql';
 import { CreateAmenityOptionInput } from './dtos/create-amenity-option.input';
 import { UpdateAmenityOptionInput } from './dtos/update-amenity-option.input';
+import { ForbiddenException, UseGuards } from '@nestjs/common';
+import { AuthGuard } from '@src/guards/auth.guard';
+import { UserType } from '@prisma/client';
 
+@UseGuards(AuthGuard)
 @Resolver(() => AmenityOption)
 export class AmenityOptionResolver {
   constructor(private readonly amenityOptionService: AmenityOptionService) {}
@@ -13,8 +21,15 @@ export class AmenityOptionResolver {
     return this.amenityOptionService.getAmenityOptionById(id);
   }
 
+  // @UseGuards(AuthGuard)
   @Query(() => AmenityOptionList, { name: 'amenityOptions' })
-  async getAllAmenityOptions() {
+  async getAllAmenityOptions(@Context() ctx: CtxType) {
+    const userType = ctx.user.userType;
+    if (userType !== UserType.SUPERADMIN) {
+      throw new ForbiddenException(
+        'You are not authorized to access this resource',
+      );
+    }
     return this.amenityOptionService.getAll();
   }
 
@@ -38,10 +53,18 @@ export class AmenityOptionResolver {
     );
   }
 
+  // @UseGuards(AuthGuard)
   @Mutation(() => AmenityOption)
   async deleteAmenityOption(
+    @Context() ctx: CtxType,
     @Args('amenityOptionId', { type: () => Int }) amenityOptionId: number,
   ) {
+    const userType = ctx.user.userType;
+    if (userType !== UserType.SUPERADMIN) {
+      throw new ForbiddenException(
+        'You are not authorized to access this resource',
+      );
+    }
     return this.amenityOptionService.delete(amenityOptionId);
   }
 }
